Obtener una lista de los conjuntos de datos disponibles de DataFaro

Usage

list_datasets(
  dataset = NULL,
  indicator = NULL,
  area = NULL,
  institution = NULL,
  freq = NULL,
  .token = Sys.getenv("DATAFARO_TOKEN")
)

Arguments

dataset

character Código del conjunto de datos (opcional)

indicator

character Código del indicador (opcional)

area

character Código del área (opcional)

institution

character Código de la institución (opcional)

freq

character Código de la frecuencia

H Horaria

D Diaria

W Semanal

M Mensual

Y Anual

.token

character Token de acceso a la API. Por defecto se utiliza el token guardado en la variable de entorno DATAFARO_TOKEN. Si no se ha guardado un token, se debe especificar el token manualmente.

Value

Un data.frame con la lista de conjuntos de datos. El data.frame tiene las siguientes columnas:

id Código del conjunto de datos

indicator Código del indicador

name Nombre del conjunto de datos

institution Código de la institución

description Descripción del conjunto de datos

status Estado del conjunto de datos

A Activo

I Inactivo

D Descontinuado

R Reemplazado

freq Código de la frecuencia

H Horaria

D Diaria

W Semanal

M Mensual

Y Anual

interval El número de veces por frecuencia que el conjunto de datos es actualizado

last_update Fecha de la última actualización

next_update Fecha de la próxima actualización

Examples

if (FALSE) {
list_datasets()

list_datasets(dataset = "imae")

list_datasets(area = "ECONREAL")
}